// src/app/client-view/client-insurance/models/insurance-objective.model.ts
export class InsuranceObjective {
    crmId: string;
    name: string;
    deathAdditionalExpenses: number;
    deathDebtPaid: number;
    deathIncomeReplaced: number;
    deathIncomeTimeYears: number;
    deathSelfInsurance: number;
    ipIncomeReplacedPercentage: number;
    ipRetirementContributionPercentage: number;
    ipSelfInsurance: number;
    ipWaitingPeriodMonths: number;    
    tpdAdditionalExpenses: number;
    tpdDebtPaidPercentage: number;
    tpdIncomeReplaced: number;
    tpdIncomeTimeYears: number;
    tpdSelfInsurance: number;
    traumaAdditionalExpenses: number;
    traumaDebtPaidPercentage: number;
    traumaIncomeReplaced: number;
    traumaIncomeTimeYears: number;
    traumaSelfInsurance: number;
    primaryClientId: string;
}